+++ src/main/kotlin/com/example/ggyunispring/adapter/dto/request/CreateDiaryRequestDTO.kt
@@ -1,7 +1,7 @@
 package com.example.ggyunispring.adapter.dto.request
 
-import com.example.ggyunispring.domain.enum.DiaryType
-import com.example.ggyunispring.domain.enum.Emotion
+import com.example.ggyunispring.domain.diary.DiaryType
+import com.example.ggyunispring.domain.diary.Emotion
 import org.hibernate.validator.constraints.Length
 import java.time.LocalDate
 import javax.validation.constraints.NotNull
@@ -43,4 +43,4 @@ data class CreateDiaryRequestDTO(
     @field:Length(min = 1)
     val longitude: String,
 
-)
+    )
