+++ src/main/kotlin/com/example/ggyunispring/application/LoginService.kt
@@ -1,8 +1,8 @@
 package com.example.ggyunispring.application
 
 import com.example.ggyunispring.adapter.infrastructure.jwt.JwtProvider
-import com.example.ggyunispring.domain.entity.Member
-import com.example.ggyunispring.domain.repository.MemberRepository
+import com.example.ggyunispring.domain.member.Member
+import com.example.ggyunispring.domain.member.MemberRepository
 import com.example.ggyunispring.adapter.dto.request.GoogleLoginRequestDTO
 import com.example.ggyunispring.adapter.dto.response.LoginResponseDTO
 import com.example.ggyunispring.domain.exception.member.GoogleIdTokenException
