+++ src/main/kotlin/com/example/ggyunispring/domain/CreateModifyTime.kt
@@ -1,4 +1,4 @@
-package com.example.ggyunispring.domain.entity
+package com.example.ggyunispring.domain
 
 import org.springframework.data.annotation.CreatedDate
 import org.springframework.data.annotation.LastModifiedDate
