+++ src/main/kotlin/com/example/ggyunispring/domain/diary/DiaryType.kt
@@ -1,4 +1,4 @@
-package com.example.ggyunispring.domain.enum
+package com.example.ggyunispring.domain.diary
 
 enum class DiaryType {
     YELLOW1, YELLOW2, BLUE1, BLUE2
