+++ src/main/kotlin/com/example/ggyunispring/domain/diary/Emotion.kt
@@ -1,4 +1,4 @@
-package com.example.ggyunispring.domain.enum
+package com.example.ggyunispring.domain.diary
 
 enum class Emotion {
     MAD, SAD, PANIC, NORMAL, HAPPY
